Added second backend and cleaned up config
Need to spend some time thinking about how to template adding in a new
backend. At the moment the two that I've chosen have been pretty easy
but I think that the next that I choose should be a non-interpreted
language so I get a better feel for how a general backend should be
added in.